requirements.txt
telethon==1.34.0
python-telegram-bot==20.8
